tiktoken_azure_sdk/requirements.txt
azure-ai-generative[evaluate]
